legacy_v1/requirements.txt
yfinance
pandas
numpy
requests
pyarrow
tqdm
nselib
pandas_market_calendars
google-generativeai
streamlit
matplotlib
seaborn